changed back the INI_USRDEF_WID to 2^18
It's because the bit-width of m_wid in CThreadSlm is still 18,
so that the system static language model could only support the
wordID range of [0, 2^18-1].

Later, we may reduce the BITS_BOW to 12, and BITS_PR to 14, so
that the static SLM could support upto 2^22 words.

And we need to investigate if the current float number compress
algorithm is efficient, to see if we could use k-means++ to do the
clustering.