wrapper/ibus: port to Python 3
Code was incompletely ported to Python 3:
- `print` statement remains.
- `u''` literals remain.

Besides this, using `type()` on literals is unnecessary since Python
2.2:
https://docs.python.org/3.8/whatsnew/2.2.html#old-and-new-classes
E.g. `type(True)` can be replaced with `bool` and `type(1)` can be
replaced with `int`.

Most calls to`set(...)` can be replaced with `{...}` (except `set()` for
creation of empty set).

the patch is contributed by @Arfrever

fix #111

Signed-off-by: Kefu Chai <tchaikov@gmail.com>